requirements.txt
streamlit
google-generativeai
streamlit-option-menu
webbrowser